Split parser builder into its own project: Ramjet.
This commit includes a bunch of renaming and cleanup.
The source code file structure should now be more dub-friendly too.

The original project was 'xdc', the Cross D Compiler or eXperimental D
Compiler. Having a good parser is requisite for that work, should I ever
get back around to it, and the parser generator is useful for many other
things as well.